requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
streamlit
anthropic
openai
pytest 